requirements.txt
# Uncomment to enable Azure Monitor OpenTelemetry
# Ref: aka.ms/functions-azure-monitor-python
# azure-monitor-opentelemetry

azure-functions
azure-data-tables
azure-storage-blob>=12.0.0
azure-data-tables>=12.0.0
python-multipart
azure_functions_multipart